move mocks to /mocks
