+++ controllers/pipelines/internal/mocks/mock_service_manager.go
@@ -1,6 +1,6 @@
 //go:build unit
 
-package testutil
+package mocks
 
 import (
 	"context"
